+++ src/constants/constants.tsx
@@ -229,3 +229,35 @@ export const SUI_SOLVER_ADDRESS =
 
 export const SUI_DEFAULT_NETWORK_FEE = 0.03;
 export const BITCOIN_DEFAULT_NETWORK_FEE = 0.49;
+
+export type StarknetChain = "starknet" | "starknet_sepolia" | "starknet_devnet";
+
+export const STARKNET_CONFIG: Record<
+  StarknetChain,
+  {
+    chainId: string;
+    nodeUrl: string[];
+  }
+> = {
+  starknet: {
+    chainId: "0x534e5f4d41494e",
+    nodeUrl: [
+      "https://starknet.api.onfinality.io/public",
+      "https://starknet.drpc.org",
+      "https://1rpc.io/starknet",
+      "https://rpc.starknet.lava.build",
+    ],
+  },
+  starknet_sepolia: {
+    chainId: "0x534e5f5345504f4c4941",
+    nodeUrl: [
+      "https://starknet-sepolia-rpc.publicnode.com",
+      "https://starknet-sepolia.drpc.org",
+      "https://starknet.api.onfinality.io/public/sepolia",
+    ],
+  },
+  starknet_devnet: {
+    chainId: "",
+    nodeUrl: [""],
+  },
+};